requirements.txt
streamlit==1.45.1
huggingface_hub==0.21.3
python-dotenv==1.0.0
pandas==2.2.3 